pyproject.toml
[project]
name = "projet-cassandra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cassandra-driver>=3.29.1",
    "pyasyncore>=1.0.5",
]
